requirements-dev.txt
# Development dependencies for auto-building
watchdog>=2.1.0

# Core dependencies (if not already installed)
pyinstaller>=5.0
mss>=6.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
pyautogui>=0.9.50
keyboard>=0.13.0
pyyaml>=5.4.0
pydirectinput>=1.0.4




